fixed useFetchData disable , fixed load more button
I fixed the useFetchData and made it stop going into an infinite loop. I also rewrote it in order to return an object that has the next url to fetch and the results from the api call. I also fixed the load more button and now it works even when you return from the pokeScreen component.